requirements.txt
audioplayer==0.6
dtw==1.4.0
gtts==2.2.4
noisereduce==2.0.0
numpy==1.21.4
Pillow==9.0.1
playsound==1.3.0
pyaudio==0.2.11
requests==2.22.0
scipy==1.6.0
sonopy==0.1.2
SoundFile==0.10.3.post1
torch==1.9.1
torchcrf==1.1.0
tqdm==4.54.1
transformers==4.1.1
youtube_search==2.1.0
